


\alias{fetal.lamb}
\docType{data}
\name{fetal.lamb}
\title{Fetal lamb movements}
\Rdversion{1.0}

\description{
  Counts of fetal lamb movements in each of 240 consecutive 5-second
  intervals.  This is a classic data set for illustrating hidden Markov models.
}

\usage{data(fetal.lamb)
fetal.lamb}

\format{A numeric vector of length 240.}

\references{

  Leroux and Puterman (1992) "Maximum-Penalized-Likelihood Estimation
  for Independent and Markov-Dependent Mixture Models,"
  \emph{Biometrics} (48) 545 -- 558.

}

\keyword{datasets}
