\name{simulate.clickstream}
\alias{SimulateFakeClickstreamParams}
\alias{SimulateClickstream}
\title{Simulating clickstream parameters and data}
\description{Simulate fake data (or parameters) for testing the nested
  HMM clickstream model.}
\usage{

SimulateFakeClickstreamParams(S2, S1, S0)
SimulateClickstream(number.of.streams, params = NULL, mean.session.size, S2, S1, S0)

}

\arguments{

\item{number.of.streams}{The number of clickstreams to simulate.
  (I.e. the number of 'users'.)}

\item{ params}{ If non-\code{NULL}, A list containing initial
  distributions and transition probability matrices for the nested HMM.
  If \code{NULL} then params will be generated by a call to
  \code{SimulateFakeClickstreamParams}.}

\item{mean.session.size}{The expected number of sessions per clickstream.}

\item{S2}{The dimension of the session-level hidden Markov chain.
  Only used if params is \code{NULL}.}

\item{S1}{The dimension of the click-level hidden Markov chain.
  Only used if params is \code{NULL}.}

\item{S0}{The dimension (number of unique values) in the observed
  data.  Only used if params is \code{NULL}.}

}

\value{ \code{SimulateFakeClickstreamParams} produces a list containing
  initial distributions and transition probability matrices for a model
  of the specified dimensions.  Each probability distribution is sampled
  from the uniform Dirichlet distribution of the appropriate dimension.

  \code{SimulateClickstream}
  A list of 2 elements.
  \itemize{
    \item{ A list contaiing \code{number.of.streams} clickstreams.}
    \item{ A list containing the true parameter values used in the
       simulation.  If params is non-\code{NULL} then this is \code{params}.
       Otherwise it is the set of simulated parameters.}
     }
}
