\name{summary.dirichlet.process.mvn}
\title{Summarize a DirichletProcessMvn Model}

\alias{summary.DirichletProcessMvn}

\description{S3 method to summarize a \code{\link{DirichletProcessMvn}}
  model object}

\usage{
  \method{summary}{DirichletProcessMvn}(object, burn = NULL, ...)
}

\arguments{
  \item{object}{An object of class \code{\link{DirichletProcessMvn}}.}
  \item{burn}{The number of MCMC iterations to discard as burn-in.}
  \item{\dots}{Extra named arguments are ignored.}
}
